refaktoring

+++ api_v2/views.py
@@ -50,7 +50,7 @@ def getGame_exp(request,id):
     cursor = connect().cursor()
     raw_query = '''SELECT players.id,COALESCE(nick,'unknown') as player_nick,
     localized_name as hero_localized_name,
-    ROUND((matches.duration::numeric / 60),2)::float as match_duration_minutes,
+    ROUND((matches.duration::numeric / 60),2) as match_duration_minutes,
     COALESCE(xp_hero,0) + COALESCE(xp_creep,0)+ COALESCE(xp_other,0) + COALESCE(xp_roshan,0) as experiences_gained,
     level as level_gained,
     matches.radiant_win = (player_slot BETWEEN 0 and 4) as winner,
